+++ R/evalOptimizationState.R
@@ -3,7 +3,7 @@
 # logs point and results
 # return y-value(s), exectime, and potential erorr msg
 evalOptimizationState = function(learner, task, resampling, measures, par.set, bits.to.features, control,
-  opt.path, show.info, dob, state, remove.nas) {
+  opt.path, show.info, dob, state, remove.nas, resample.fun) {
 
   setSlaveOptions()
   y = setNames(rep(NA_real_, length(measures)), vcapply(measures, measureAggrName))
@@ -39,7 +39,7 @@ evalOptimizationState = function(learner, task, resampling, measures, par.set, b
       state, NA_real_, remove.nas, stage = 1L)
   if (set.pars.ok) {
     exec.time = measureTime({
-      r = resample(learner2, task, resampling, measures = measures, show.info = FALSE)
+      r = resample.fun(learner2, task, resampling, measures = measures, show.info = FALSE)
     })
 
     if (control$tune.threshold) {
@@ -81,7 +81,7 @@ evalOptimizationState = function(learner, task, resampling, measures, par.set, b
 # adds points to path
 # returns list of lists, the single eval results
 evalOptimizationStates = function(learner, task, resampling, measures, par.set, bits.to.features, control,
-  opt.path, show.info, states, dobs, eols, remove.nas, level) {
+  opt.path, show.info, states, dobs, eols, remove.nas, resample.fun, level) {
 
   n = length(states)
   if (length(dobs) == 1L)
@@ -93,7 +93,8 @@ evalOptimizationStates = function(learner, task, resampling, measures, par.set, 
   res.list = parallelMap(evalOptimizationState, dobs, states, level = level,
     more.args = list(learner = learner, task = task, resampling = resampling,
       measures = measures, par.set = par.set, bits.to.features = bits.to.features,
-      control = control, opt.path = opt.path, show.info = show.info, remove.nas = remove.nas))
+      control = control, opt.path = opt.path, show.info = show.info, remove.nas = remove.nas,
+      resample.fun = resample.fun))
   # add stuff to opt.path
   for (i in seq_len(n)) {
     res = res.list[[i]]
@@ -106,15 +107,15 @@ evalOptimizationStates = function(learner, task, resampling, measures, par.set, 
 }
 
 evalOptimizationStatesTune = function(learner, task, resampling, measures, par.set, control,
-  opt.path, show.info, states, dobs, eols, remove.nas) {
+  opt.path, show.info, states, dobs, eols, remove.nas, resample.fun) {
 
   evalOptimizationStates(learner, task, resampling, measures, par.set, NULL, control,
-    opt.path, show.info, states, dobs, eols, remove.nas, "mlr.tuneParams")
+    opt.path, show.info, states, dobs, eols, remove.nas, resample.fun, "mlr.tuneParams")
 }
 
 evalOptimizationStatesFeatSel = function(learner, task, resampling, measures, bits.to.features, control,
   opt.path, show.info, states, dobs, eols) {
 
   evalOptimizationStates(learner, task, resampling, measures, NULL, bits.to.features, control,
-    opt.path, show.info, states, dobs, eols, FALSE, "mlr.selectFeatures")
+    opt.path, show.info, states, dobs, eols, FALSE, resample, "mlr.selectFeatures")
 }