+++ R/tuneMultiCritNSGA2.R
@@ -1,4 +1,4 @@
-tuneMultiCritNSGA2 = function(learner, task, resampling, measures, par.set, control, opt.path, show.info) {
+tuneMultiCritNSGA2 = function(learner, task, resampling, measures, par.set, control, opt.path, show.info, resample.fun) {
 
   requirePackages("mco", why = "tuneMultiCritNSGA2", default.method = "load")
   low = getLower(par.set)
@@ -9,7 +9,7 @@ tuneMultiCritNSGA2 = function(learner, task, resampling, measures, par.set, cont
     lower.bounds = low, upper.bounds = upp,
     learner = learner, task = task, resampling = resampling, measures = measures,
     par.set = par.set, ctrl = control, opt.path = opt.path, show.info = show.info,
-    convertx = convertXNumeric, remove.nas = FALSE)
+    convertx = convertXNumeric, remove.nas = FALSE, resample.fun = resample.fun)
   args = c(args, control$extra.args)
 
   do.call(mco::nsga2, args)