+++ R/tuneParams.R
@@ -26,6 +26,10 @@
 #' @param control [\code{\link{TuneControl}}]\cr
 #'   Control object for search method. Also selects the optimization algorithm for tuning.
 #' @template arg_showinfo
+#' @param resample.fun [\code{closure}]\cr
+#'   The function to use for resampling. Defaults to \code{\link{resample}} and should take the
+#'   same arguments as, and return the same result type as, \code{\link{resample}}. This
+#'   parameter must be the default when \code{mbo} tuning is performed.
 #' @return [\code{\link{TuneResult}}].
 #' @family tune
 #' @note If you would like to include results from the training data set, make
@@ -76,12 +80,13 @@
 #' print(head(as.data.frame(res$opt.path)))
 #' }
 #' @seealso \code{\link{generateHyperParsEffectData}}
-tuneParams = function(learner, task, resampling, measures, par.set, control, show.info = getMlrOption("show.info")) {
+tuneParams = function(learner, task, resampling, measures, par.set, control, show.info = getMlrOption("show.info"), resample.fun = resample) {
   learner = checkLearner(learner)
   assertClass(task, classes = "Task")
   measures = checkMeasures(measures, learner)
   assertClass(par.set, classes = "ParamSet")
   assertClass(control, classes = "TuneControl")
+  assertFunction(resample.fun)
   if (!inherits(resampling, "ResampleDesc") &&  !inherits(resampling, "ResampleInstance"))
     stop("Argument resampling must be of class ResampleDesc or ResampleInstance!")
   if (inherits(resampling, "ResampleDesc") && control$same.resampling.instance)
@@ -101,6 +106,10 @@ tuneParams = function(learner, task, resampling, measures, par.set, control, sho
     TuneControlIrace = tuneIrace,
     stopf("Tuning algorithm for '%s' does not exist!", cl)
   )
+
+  if (cl == "TuneControlMBO" && !identical(resample.fun, resample)) {
+    stop("resample.fun must be the default value when mbo tuning is performed.")
+  } 
   opt.path = makeOptPathDFFromMeasures(par.set, measures, include.extra = (control$tune.threshold))
   if (show.info) {
     messagef("[Tune] Started tuning learner %s for parameter set:", learner$id)
@@ -108,7 +117,7 @@ tuneParams = function(learner, task, resampling, measures, par.set, control, sho
     messagef("With control class: %s", cl)
     messagef("Imputation value: %g", control$impute.val)
   }
-  or = sel.func(learner, task, resampling, measures, par.set, control, opt.path, show.info)
+  or = sel.func(learner, task, resampling, measures, par.set, control, opt.path, show.info, resample.fun)
   if (show.info)
     messagef("[Tune] Result: %s : %s", paramValueToString(par.set, or$x), perfsToString(or$y))
   return(or)